+++ src/clickgen/packer/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from clickgen.packer.windows import pack_win
+from clickgen.packer.x11 import pack_x11
+
+__all__ = ["pack_win", "pack_x11"]