+++ src/clickgen/writer/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from clickgen.writer.windows import to_cur, to_win
+from clickgen.writer.x11 import to_x11
+
+__all__ = ["to_x11", "to_cur", "to_win"]