+++ CHANGELOG.md
@@ -9,11 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- refactor: removed old source code
+- add: python 3.7 support
+- add: **Windows** and **macOS** support #24
+- init: `cursor`, `configparser`, `packer`, `parser` and `writer` module
 - 'Twitter' and 'Download' links added on PYPI page
+- cli init ( `clickgen -h` )
 
 ### Changed
 
 - Changed docs theme to 'furo'
+- chore: updated template variables inside `packer.windows`
+- make: install all dependencies with `make install_deps` command
+- chore: directory renamed `examples` -> `samples`
 
 ## [1.2.0] - 26 Mar 2022